Integrated AlertCards with FormatAlert
